hostapd: Disable Hosapd interface on STA Disconnect
Currently when the STA in the STA+AP interface disconnects from ROOT AP,
wpa_supplicant sends STOP_AP command to hostap interface to stop AP
operations. But the AP interface is not completely disabled here,
due to which we could receive some packets to this interface when the
STA is performing active scan. Hence use "DISABLE" command to disable
the AP interface.

Change-Id: If010f7a5e1877d0af5496327f70d24ce939d4053
Signed-off-by: Sriram R <srirrama@codeaurora.org>