busybox: prevent compile hang with bzip2 enabled
The BZIP2_SMALL option was not being exposed via Config.in which
caused the build to fail as 'yes' is piped to the config during
build.  As it's expecting a number, it gets stuck in a loop.

Signed-off-by: Rob Mosher <nyt-openwrt@countercultured.net>
(cherry picked from commit 8a3582fa8ae9a894a299ef28d0aceac1a8c4272b)

Change-Id: I45730166b5504f2f89cc22c79129522773b99be9
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>