openvpn: add list element parsing.
For the parameters tls-cipher and ncp-ciphers more than one option can
be used in the OpenVPN configuration, separated by a colon, which should
be implemented as a list in order to configure it more clearly. By
adding the new OPENVPN_LIST option to the openvpn.options file with the
tls-cipher and ncp-cipher parameters, uci can now add this option as a
"list" and the init script will generate the appropriate OpenVPN
configuration from it.

Signed-off-by: Hans Dedecker <dedeckeh@gmail.com>
Signed-off-by: Florian Eckert <fe@dev.tdt.de>
(cherry picked from commit 675eb747aa1cb76d4d6e715728592974acac8cc1)

Change-Id: I6b80609f51d5981ff1c55cf91ec2a15c29712869
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>