openvpn: update to 2.4.6.
Update the OpenVPN package to version 2.4.6, refresh patches and drop
menuconfig options which are not supported upstream anymore.

Also fix the x509-alt-username configure flag - it is not supported
by mbedtls and was syntactically wrong in the Makefile - and the
port-share option which has been present in menuconfig but not been
used in the Makefile.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>
(cherry picked from commit 3082370551e78195def4b3989eba60bc6149cf6d)

Change-Id: I950efddab6914f92e956300f2dd0336dab0c61a5
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>