busybox: handle crypt() errors in loginutils
The crypt(3) function is allowed to fail with either EINVAL or ENOSYS when
the given salt is either invalid or when the requested algorithm is not
implemented.

In such a case, libbb's pw_encrypt() function will silently convert the
crypt() NULL return value into an empty string which is then processed
without further errors by utilities such as chpasswd or passwd, causing
them to set an empty password when an unsupported cipher is requested.

Patch the relevant users of pw_encrypt() to abort in case an empty hash
is returned by pw_encrypt() in order to mitigate the problem.

Signed-off-by: Jo-Philipp Wich <jo@mein.io>
(cherry picked from commit 1211832977b98c491d1198ab66c4f8ffc0886a87)

Change-Id: Ifa2909e268a614a19baff6e0494b042d0cbcc9fa
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>