Do not disassociate not-associated STA on timeout
If the ap_handle_timer() timeout is reached for a not-associated STA, do
not default to disassociating that STA first since Disassociation frame
is not really appropriate to send to a STA that is not in associated
state. Instead, skip directly to deauthentication and STA entry removal.

Taken from the below upstream hostapd commit
https://w1.fi/cgit/hostap/commit/?id=f3a841bbff689cf4e89831afe3192ff4d9

Without this fix, there are lots of Disassociation frame when run mdk3
authentication DoS test, which will cause FW management TX failed, as
there is no enough TX buffer. The logs as below,
      [ 1437.614375] [phy2] FWLOG: [1333126] vap-0 WAL_DBGID_MGMT_TX_FAIL \
	      ( 0xbadba12, 0x2f, 0x2, 0x0, 0xa0 )
      [ 1437.622578] [phy2] FWLOG: [1333128] vap-0 WAL_DBGID_MGMT_TX_FAIL \
	      ( 0xbadba12, 0x2f, 0x2, 0x0, 0xa0 )

Change-Id: Id0d1e651b84d45e58ab0ba56a6d84915983d8b4d
Signed-off-by: Jouni Malinen<jouni@codeaurora.org>
Signed-off-by: Miaoqing Pan <miaoqing@codeaurora.org>