libnl:fix the lock issue in rtnl_link_af_unregister api
The lock that is aquired in the function rtnl_link_af_unregister is not unlocked.
Also, while returning on empty ops, its acquiring a lock instead of returning.
Due to the above code, dependent applications never exit.
This patch fixes the issue by returning error on empty ops argument.
It also handles the lock unlock properly.

Change-Id: I5d8415d1ed70541c8647fcd97db0500d05f9ed24
Signed-off-by: Sirisha Mosalikanti <sirisham@codeaurora.org>