MIPS: AP151: Modify ddr cas latency to 5
Modify the DDR Cas latency to 5 for AP151.
Earlier it was not defined in Uboot config and a default
of 4 was taken. There was a customer request to modify
the ddr latency to 5. Also note that the LSDK version
used 5 for the latency.

The change is for the boot package used along
with 3.3.8 kernel version.

Change-Id: I68b35bd316fdf476cba882f40415de8c46fe3a7d
Signed-off-by: Charanya Venkatraman <charanya@codeaurora.org>