configs: ipq6018: Increase CMA Size to 32 MB
This patch increases the CMA Size from 16 MB
to 32 MB.

When we have the CMA Size as 16 MB, dma alloc
coherent allocations were done from this pool
because of which it was failing when NSS was
trying to allocate memory using these DMA APIs.

This patch should later be removed as this is
just a temporary fix till we find a permanent
solution.

Change-Id: I45c992411dad2c1bbfc4238aa3542348de916f80
Signed-off-by: speriaka <speriaka@codeaurora.org>