base-files: sysupgrade: fix in platform_copy_config
while copying the config, for nand and emmc flash types, a direct mount
attempt is done to the /tmp/overlay folder without checking if it
even exists.
If /tmp/overlay is not present, copy fails and the configs won't be
saved.
This change creates the /tmp/overlay directory if it doesn't exist
before attempting to mount the ext4 or ubifs for emmc and nand
respectively.

Change-Id: If21ac918085b6155a58be4e5423ae8e186ee577e
Signed-off-by: Sirisha Mosalikanti <sirisham@codeaurora.org>