sysupgrade: Return error for verification failure
sysupgrade calls platform_check_image to authenticate image
before flashing. This API returns 0 on success but does not
return anything on failure. So caller continues executing.
This results in two-time authentication of the image during
failure.

This change adds code to return error on failure. So caller
ceases to execute further.

Change-Id: I863551aae72fa7072b6edac687409b2f44f361dd
Signed-off-by: Pradeep Das <pkdas@codeaurora.org>