download.mk : add a variable to skip download from CAF mirror
If we need to skip the download from CAF mirror for git projects,
we can set this variable to true in a package Makefile.

Why we introduced this?
The proprietary sigma-dut and open source sigma-dut-open packages
have same git repo names. So, during the CAF mirror check, a clash
will occur.
proprietary sigma-dut is already hosted in github.qualcomm.com which
is as good as CAF. So, we can now create CAF mirror for open source
sigma-dut-open package.

Change-Id: Iffe3122739b2f76644a26d30701feaf66fbd3cf2
Signed-off-by: Sirisha Mosalikanti <sirisham@codeaurora.org>