hostapd: Fix to downgrade bandwidth in radar detection
Currently upon radar detection we search channels on same bandwidth.
This way we may not get any other channels if there are no channels
available on that bandwidth, then AP service will be stopped. To avoid
this problem this patch search the channels by downgrading the bandwidth,
this scenario applicable during CAC as well.
Also, in normal case hostapd_disable_iface and hostapd_enable_iface will be
done while switching to another dfs channel upon radar detection. In certain
scenario radar detected event can come during hostapd_disable_iface is in progress
and iface->current_mode will be NULL in that scenario. Currently we are ignoring
this scenario and proceeding with the radar detection logic, this will trigger
segmentation fault. To fix this, avoid proceeding the radar detection event
if the iface->currentmode is NULL.

Change-Id: Ifb8dbfbc2bf6da6cdc82698b5e4d62f4cbd7c760
Signed-off-by: Seevalamuthu Mariappan <seevalam@codeaurora.org>