hostapd: set support channel width flag when switching to 40Mhz
When utilizing chan_switch cmd to change bw to 40 MHz when AP is
brought up in 20 MHz, support channel width flag is unset in
beacon/probe response resulting in STA to downgrade to 20 MHz.
In addition, secondary offset is set to -1/1 depending if new mode
is HE40+ or HE40-. Corrected to reflect support for 40 MHz in HT
capabilities and secondary channel offset to reflect new 40MHz
setting.

Change-Id: Idc441640626de6dbf02803fe2ae5fa9bf10b179b
Signed-off-by: Muna Sinada <msinada@codeaurora.org>