hostapd: fix FT-EAP STA re-association failure
When FT-EAP Over-the-air STA tries to re-associate back with the
previous association not terminated, we use the previous aid and
listen_interval in the SET_STATION for the STA in WLAN_STA_ASSOC
state.

This will fail in cfg80211_check_station_change() for the AID,
listen interval not to be set for the STA in WLAN_STA_ASSOC
state.

Fix this by doing hostapd_drv_sta_remove() to proceed with
NEW_STATION instead of SET_STATION.

Signed-off-by: Sathishkumar Muruganandam <murugana@codeaurora.org>
Change-Id: Ie9951362b6a6b2f004802382c4bd4e88d735b499