busybox: ipq customizations
1. Enable busybox tftp for file transfers

tftp is simple utility which allows the client to get
from or put a file onto a remote host.
We are enabling busybox tftp for optimizing the image size.
Usages:
 *Get a file from remote host:
  tftp -g -r filename.txt 192.168.0.1
 *Put a file to remote host:
  tftp -p -l filename.txt 192.168.0.1

2. install basic inittab if busybox setting on
3. add opkg host tools for installation

This is normally pulled in by base-files, but if base-files is not
selected we can't create opkg's and an image, so let's add the
depends here too so image creation works on super minimal openwrt
busybox based builds

4. add a few more init bits for standalone busybox
5. config: turn on third status line for dd

This is used for performance measurements, so let's add it
by default

6. init: mount debugfs if using bb init
7. add some more default initialization

- default password
- udhcpc event hander script
- start watchdog

8. load all modules in extra init script
9. swap udev for mdev in minimal image

    kernel hotplug uses netlink for messages and mdev does not
    completly work right
10. Enable customization & FTP
11. Create inetd.conf from root etc folder
12. Enable IPCS command
13. Enable config to store output of history

This Change is to facilitate the request to capture output
of history command as part of crashdumps, to do that we need
to capture output of history command in a file first,this
change enables that and captures history in .ash_history file

14. Enable STAT and STAT_FORMAT as default
15. Enable usleep in busybox for audio
16. Enable TASKSET in busybox for audio
17. Enable run-parts as default config
18. Install inetd
19. Change default max len in busybox to 1024

This patch changes the default max length supported
in command line for busybox from 512 to 1024

20. Enable ftpd write and authentication

These 2 configs are disabled as part of busybox upgrade.

Change-Id: I2e547916da1c11102ae0c9c551bb699bafbaedc9
Signed-off-by: pavir <pavir@codeaurora.org>
Signed-off-by: Avinash Pandey <avinasv@codeaurora.org>
Signed-off-by: Prem Thilagar <premt@codeaurora.org>
Signed-off-by: Anto Norbert <norbrt@codeaurora.org>
Signed-off-by: Naresh Kumar Mehta <naresh@codeaurora.org>
Signed-off-by: Ram Chandra Jangir <rjangi@codeaurora.org>
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>
Signed-off-by: Sirisha Mosalikanti <sirisham@codeaurora.org>