gen-relreport: add customizations for 19.07 (fig)
The metadata has changed in fig as below:
1. source is now a hash type
2. subdir key field of package is available as path key field in
source hash.
Relevant canges have been done to address the below:
We need 1. to publish the package name.
We need 2. to publish the feed name.

To fix the 'circular data structure error' while generating
the xml output, the package->{src} had to be overwritten by
package->{src}->{name}

Change-Id: Iec18cf048311d4b47ddc48c9eec429c893ea1c71
Signed-off-by: Sirisha Mosalikanti <sirisham@codeaurora.org>