ipq807x: config-5.4: usb: enable USB_GADGET config
Enabled CONFIG_USB_GADGET as y so that CONFIG_USB_BAM
can also be made as built-in. If USB_GADGET config is
enabled as module then CONFIG_USB_BAM will also be
enabled as module, due to which compilation errors are
observed in coresight files and usb_bam package needs
to be defined to clear the dependencies with usb-f-qdss
package with usb_bam module.

Signed-off-by: Balaji Prakash J <bjagadee@codeaurora.org>
Change-Id: Ie056c461e96ff1a3d48aa0d80788c3b9fbfedd0d