musl: Fix for CVE-2016-8859.
fix free of uninitialized buffer pointer on error in regexec
the fix in commit c3edc06d1e1360f3570db9155d6b318ae0d0f0f7 for
CVE-2016-8859 used gotos to exit on overflow conditions, but the code
in that error path assumed the buffer pointer was valid or null. thus,
the conditions which previously led to under-allocation and buffer
overflow could instead lead to an invalid pointer being passed to
free.

Change-Id: Ia7eba2b32ba0b301775860a5d5a1504e09c5ccfe
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>