musl: Fix for CVE-2017-15650.
in dns parsing callback, enforce MAXADDRS to preclude overflow
MAXADDRS was chosen not to need enforcement, but the logic used to
compute it assumes the answers received match the RR types of the
queries. specifically, it assumes that only one replu contains A
record answers. if the replies to both the A and the AAAA query have
their answer sections filled with A records, MAXADDRS can be exceeded
and clobber the stack of the calling function.

this bug was found and reported by Felix Wilhelm.

Change-Id: I6cc8dc7e6ae0bc3f1db1c42c76224cb65087554b
Signed-off-by: Sohan Madhav Bangaru <sbangaru@codeaurora.org>