include: Add support to use qsdk-package
To use the patches or changes for a package, the patches directory
under the qsdk-package directory, must has the same path as the base
package resource directory relative to the qsdk project root.

For example, if we want to use the patches under the base package directory:

packages/libs/libnl/patches we need to create the directory:

<qsdk-package>/packages/libs/libnl/patches and in the directory, put the
patch file with the same path and patch file names with same format as
where they are defined in the base package.

Change-Id: I8b2bc491ed8ecc05e5351a5066464d44f076c359
Signed-off-by: Ram Chandra Jangir <rjangir@codeaurora.org>