qsdk.mk: Enable iperf for Premium profile
The commit 34b6c8b0759f2ae68afb2953280e27f23b9d7e5e
removes the support for single threaded variant of iperf,
the variant iperf-mt is obsolete w.r.t to 19.07.

The package "iperf" needs to be enabled in 19.07
instead of iperf-mt (as in cc/date). This will make the
iperf binary available in the image at /usr/bin

Signed-off-by: Sandhya Ks <sandhyks@codeaurora.org>
Change-Id: Iaa40cfb399dd076b663eb35482e24b3e72c5f3ea