base-files: Add ubus process to list of essential services
When we have three wireless radio cards on board, we observed
that system is somewhat slow and may alter the command sequence.

During sysupgrade we send ubus signal to procd to handle upgrade
and starts killing the running processes by expecting that ubus
signal is delivered but due to system load, process killing is
triggered and it kill's the ubus command too which was triggered
to send signal and it results in an unexpected behavior.

In order to fix this problem, we makes ubus and sh process to the
part of essential list of services which should not be SIGTERM during
sysupgrade.

Change-Id: I5c5f2af6d59cadaf8e9d00f0e626c220fae57e17
Signed-off-by: Ram Chandra Jangir <rjangir@codeaurora.org>