gen-relreport.pl: Add a sheet with tarball source names
In the release report, add a sheet which contains the tarball names
that can be downloaded from external sites. This information can be
directly used in release notes.
We do not publish the source tar balls which are from internal servers.
This is done by using the Category in the package Makefile.

Change-Id: I9d6333f1e44654a207d13338755ef53c51ad80ff
Signed-off-by: Sirisha Mosalikanti <sirisham@codeaurora.org>