ipq: Add support for sysupgrade to OpenWrt v19.07
In eMMC boot, existing qsdk versions are using rootfs_data
partition for overlayfs, but the newer OpenWrt upgrades are
using emmc overlay partition as loop device, this loop device
is associated from the unused free space of rootfs partition.

So our older qsdk versions should understand this loop device
feature and use it when we upgrade from cc/date to v19.07.
This will help to retain the user configurations in sysupgrade.

Change-Id: I50afeb1fdbc1e4ea3f5cccedaf212294c51a6abf
Signed-off-by: Ram Chandra Jangir <rjangir@codeaurora.org>