base-files: Correct the JSON file for initramfs network configs
The Openwrt 19.07 generates the /etc/config/network based on
board.json file.

In case of initramfs boot, the configs from
target/linux/ipq807x/base-files/etc/board.d/01_network are
being dumped in to /tmp/board.json - which is used in preinit,
whereas the /etc/config/network is generated from /etc/board.json.
Due to this, /tmp/board.json holds the ipq807x network configs and
the /etc/board.json holds the default network configs as in
package/base-files/files/etc/board.d/99-default_network.
The /etc/config/network which is based on /etc.board.json
also holds only the default configs.

Hence the pings fail when the DUT boots up with the
initramfs image.

Correcting the JSON path with the CFG variable in the network
configuration file, ensures that the json data is dumped in to the
correct file, and the /etc/config/network is generated properly.

Change-Id: I910cee13bd219142e28328d97ab275ecd620033f
Signed-off-by: Sandhya KS <sandhyks@codeaurora.org>