clean build: Clean compile the toolchain headers
Currently 'make clean' does not clean the toolchain kernel headers.
As a result, if there are changes made to the headers files,
these are not reflected in the staging directory during compilation
even after cleaning the build.

This change ensures that running the 'make clean' command,
will also clean the toolchain headers.

The kernel headers will get recompiled during the full build
before the target compilation, so that the updated headers
are reflected in the staging directory.

Change-Id: Ib9c619e807a7d045f6a984be826bc4c108bd9f88
Signed-off-by: Sandhya KS <sandhyks@codeaurora.org>