gen-relreport: use default profile for x86_basic.config
To generate release report accurately, we are relying on the input
parameters for target, subtarget info and in the config file for
profile info. If we are not getting the Profile info, we skip and
do not generate the output file.
However, for certain cases, like x86_basic and host component builds,
the output file with default configs is expected.
This change takes care of the same.

Signed-off-by: Sirisha Mosalikanti <sirisham@codeaurora.org>
Change-Id: I05885003026719ab6df7c81da242942ad3efae3f